move setcountdown status to different file
